+++ src/main/java/org/example/kihelp_back/teacher/dto/TeacherRequest.java
@@ -7,6 +7,6 @@ import static org.example.kihelp_back.teacher.util.ErrorMessage.TEACHER_NAME_NOT
 public record TeacherRequest(
         @NotBlank(message = TEACHER_NAME_NOT_BLANK)
         String name,
-        Integer subjectId
+        Long subjectId
 ) {
 }
