+++ src/main/java/org/example/kihelp_back/teacher/dto/TeacherResponse.java
@@ -3,6 +3,6 @@ package org.example.kihelp_back.teacher.dto;
 public record TeacherResponse(
         Integer id,
         String name,
-        Integer subjectId
+        Long subjectId
 ) {
 }
