+++ src/main/java/org/example/kihelp_back/teacher/repository/TeacherRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface TeacherRepository extends JpaRepository<Teacher, Integer> {
-    boolean existsByNameAndSubjectId(String name, int subjectId);
-    List<Teacher> findTeacherBySubjectId(int subjectId);
+    boolean existsByNameAndSubjectId(String name, Long subjectId);
+    List<Teacher> findTeacherBySubjectId(Long subjectId);
 }
